feat(skipcpio): move cpio on-disk format into separate header
Preparation for reusing cpio archive definitions.
skipcpio's PROGRAM_VERSION_STRING is currently unused so drop it.
